move utils to project root
