+++ countyapi/api.py
@@ -14,7 +14,7 @@ from tastypie.authorization import Authorization
 
 from countyapi.models import CountyInmate, CourtLocation, CourtDate, HousingLocation, HousingHistory, \
     DailyPopulationCounts, DailyBookingsCounts, ChargesHistory
-from countyapi.utils import convert_to_int
+from utils import convert_to_int
 
 
 COUNTY_API_INMATE_RESOURCE = 'countyapi.api.CountyInmateResource'
